Separating user input from interactive plotting handling, adding in interactive, snapshot, and animation .py scripts, with the generate_interactive.py core code renamed to interactive_handler.py
