+++ scripts/light_ions_plot.py
@@ -17,7 +17,7 @@ from QoL import set_highlighted_excepthook
 
 CHARGE_DIFFERENCE = False # Set True if want ionised atoms' trace to start from origin
 PLOT_DERIVATIVE = False # Plot the rate of avg charge gain
-PLOT_MODE = 2  # 0: plot all charges, 1: plot element total charges # 2: plot orbital charges
+PLOT_MODE = 1  # 0: plot all charges, 1: plot element total charges # 2: plot orbital charges
 YLIM = [0,None]
 FIGWIDTH = 3.49751*2/3
 FIGHEIGHT = 3.49751/2
